chore!: install depenedencies automatically via CMake
This change ports the QMake build to CMake - a recommended way to
build plugins according to the Qt Documentation:
https://doc.qt.io/qtcreator-extending/first-plugin.html#cmake-project

This also adds a number of useful features, for example, an ability
to fetch neovim-qt and qtcreator automatically.

Neovim-qt is downloaded and built with the Message Pack using
FetchContent CMake feature. In case of QtCreator only the
precompiled binaries are downloaded to save time.