chore: unify QtCreator fetching code with CI
Before this change, QtCreator fetching code from external directory was
different from the one from CI scripts. You'd have to specify QtCreator
version in two places and the `external` fetch was more complex,
requiring you to change file names and checksums for every platform.

This change unifies the way to download QtCreator, while simplifying
it.

Some drawbacks:
- Qt Creator now unpacks to `external` source directory, requiring you
to delete the `dist` dir, if you want a full clean build. In the most
cases, since the downloaded files are binaries, that are stable, you
would not want to do that often.
- Since Qt Packaging script is invoking CMake and comes from QtCreator
distribution, we have to download the QtCreator outside of the CMake
build process, which makes things more complex. But we only need to
do this once, in the CI, so local builds are as simple, as they were.