refactor: always use UTF8 for string decode/encode
Neovim nowadays always use UTF8 for RPC (`:h encoding`).
This change reduces the dependency on neovim-qt for that
particular thing and uses Qt built-in mechanism instead.

neovim-qt now uses internally QTextCodec internally to
decode/encode strings, but in Qt6 it was removed, so
they planned[^1] to ditch it and force UTF8 anyway,
because of the same reason.

[^1]: See one of the points here:
https://github.com/equalsraf/neovim-qt/issues/839